feat(): Refactor proposal and credit card models to use Money type for amounts. Refined some API endpoints.

+++ src/Services/CreditCard/CreditCard.API/CreditCard/RequestCreditCard/CreditCardRequest.cs
@@ -4,6 +4,6 @@ public class CreditCardRequest
 {
     public required Guid ClientId { get; set; }
     public required Guid ProposalId { get; set; }
-    public required int ExpensesLimit { get; set; }
+    public required decimal ExpensesLimit { get; set; }
     public required string CardProvider { get; set; }
 }
